SP-376 refactor: 카카오에서 받아오는 리프레시 토큰 만료시간 자료형 변경
- 카카오 로그인 Docs와 일치
  - Long -> Integer

+++ src/main/java/com/ludo/study/studymatchingplatform/auth/service/kakao/dto/KakaoOAuthToken.java
@@ -19,6 +19,6 @@ public class KakaoOAuthToken {
 	private String refreshToken;
 
 	@JsonProperty("expires_in")
-	private Long expiresIn;
+	private Integer expiresIn;
 
 }
